refactor: extract bass pitch MIDI output
